refactor: Move INVALID_MOVE constant to separate file
