+++ packages/core.ts
@@ -6,7 +6,7 @@
  * https://opensource.org/licenses/MIT.
  */
 
-import { INVALID_MOVE } from '../src/core/reducer';
+import { INVALID_MOVE } from '../src/core/constants';
 import { ActivePlayers, TurnOrder, Stage } from '../src/core/turn-order';
 import { PlayerView } from '../src/core/player-view';
