+++ src/core/constants.ts
@@ -1,0 +1,6 @@
+/**
+ * Moves can return this when they want to indicate
+ * that the combination of arguments is illegal and
+ * the move ought to be discarded.
+ */
+export const INVALID_MOVE = 'INVALID_MOVE';
